Renamed search components to autocomplete
