+++ src/components/shared/AutoComplete/SelectedServiceHeader/CloseButton/CloseButton.js
