+++ src/components/shared/AutoComplete/SelectedServiceHeader/CloseButton/CloseButton.module.scss
