+++ src/components/shared/AutoComplete/TrainAutoComplete/TrainAutoCompleteResult/TrainAutoCompleteResult.js
