+++ src/components/shared/AutoComplete/TrainAutoComplete/TrainAutoCompleteResult/TrainAutoCompleteResult.module.scss
